perf: speed up the boosting core; default the covariance cache to "auto"
Measured end-to-end on real preprocessed scRNA-seq (20,000 cells,
latent_dim=10): 1.81x at p=2,000, 1.59x at p=3,000, 1.59x at p=8,000, and
1.90x on a batch-integrated fit with 20 covariate dummies. Stability
selection gains too -- 1.46x for iteration mode, 1.32x for subsample mode,
with peak memory during a subsample run falling from 360 MB to 6 MB.

This changes no statistics. Across eighteen configurations -- plain, all
three batch_integration_mode values, flat and per-dimension mandatory_genes,
balance_obs, split_softmax, both disentanglement methods, standardize_targets,
frozen/anchored/zero-dimension transfers and both stability modes -- the
selected gene set is identical and the reconstruction loss agrees to the
sixth decimal.

Bitwise-identical changes:

* col_norms_sq was rebuilt on every training iteration although the design
  matrix is fixed for the whole fit, allocating a full-size temporary each
  time. It is hoisted, via the new `column_norms_sq` helper so a caller
  cannot open-code a value that differs from the internal one.
* The mandatory-covariate block was rebuilt on every boosting *step* although
  it depends only on the target. Built once per target instead; the solve
  stays np.linalg.solve, since a factorization would change results in the
  last bits and the solve is O(n_mandatory^3) anyway.
* The covariance cache's NaN check re-scanned columns already verified.

Result-changing changes (see CHANGELOG):

* boosting_precompute_covcache defaults to "auto" and now applies to every
  boosting path, including iteration-mode stability selection, which ignored
  it entirely. The flag is a speed setting first: building all p columns at
  once is one compute-bound matrix product, fetching them one at a time is a
  sequence of memory-bound matrix-vector products.
* The Gram is returned Fortran-ordered, so covcache[:, j] -- the only way
  allboost reads it -- is contiguous rather than a stride of 8*p bytes.
* Starting residual correlations are formed for all latent dimensions in one
  matrix product rather than one per dimension.
* A float64 target against float32 predictors used to promote the *design
  matrix*, materializing a full float64 copy; it also silently produced a
  different answer. The target is aligned to the predictors instead.
* fit ran boosting in float32 while both stability paths ran float64 on top
  of a float32 copy they had already made. They now agree on float32, which
  removes two full-size copies of the expression matrix.
* A frozen transfer with no added dimensions no longer builds a covariance
  matrix it never reads.

Subsample-mode stability selection deliberately stays lazy: precomputing per
subsample measured 3.07x faster at p=2,000 but 0.73-0.85x slower from p=6,000
upwards, and neither a memory guard nor a p threshold separates those cases
soundly. The measurement is recorded in _stability.py next to the decision.

Results are not bit-identical to 0.1.2; a fixed seed no longer reproduces its
output exactly. The selected support does not change.

+++ src/structboost/_stability.py
@@ -303,8 +303,12 @@ def stability_selection(
     -------
     StabilitySelectionResult
     """
-    sourcemat = np.asarray(sourcemat, dtype=np.float64)
-    targetmat = np.asarray(targetmat, dtype=np.float64)
+    # Dtype is the caller's; allboost aligns the target to the predictors.
+    # Forcing float64 here used to double the memory of every BAE-driven call
+    # (the caller already holds a float32 copy) for a precision difference
+    # measured at about one gene in 380.
+    sourcemat = np.asarray(sourcemat)
+    targetmat = np.asarray(targetmat)
     n = sourcemat.shape[0]
     if targetmat.shape[0] != n:
         raise ValueError(
@@ -351,6 +355,17 @@ def stability_selection(
         idx = rng.choice(n, size=sub_n, replace=False)
         # A fresh covcache per subsample is mandatory: the Gram matrix depends on
         # the rows, so the full-data cache would silently produce wrong updates.
+        #
+        # Deliberately lazy, unlike `BAE.fit`, which precomputes the whole Gram by
+        # default. Building it per subsample was measured and rejected: on real
+        # data at n=15,000 it runs 3.07x faster at p=2,000 and 2.69x at p=3,000,
+        # but 0.73-0.85x *slower* from p=6,000 upwards. The crossover is governed
+        # by the number of distinct columns a subsample actually selects, which
+        # stays roughly flat (~110-240) as p grows, so the full p x p product
+        # stops paying for itself. A memory guard cannot separate those cases --
+        # the 800 MB Gram at p=10,000 fits comfortably and still loses -- and any
+        # p threshold would be calibrated on one dataset's support size. Left
+        # lazy, which is the safe direction at every p.
         betamat = allboost(
             sourcemat[idx],
             targetmat[idx],
